Move individual tree data to species folder, and correct typo
