+++ src/data/trees/index.js
@@ -1,7 +1,11 @@
-import quercusRobur from './quercus_robur'
-import quercusPetrea from './quercus_petrea'
+/**
+ * Export a list of the Tree objects as the default for this directory.
+ */
+
+import quercusRobur from './species/quercus_robur'
+import quercusPetraea from './species/quercus_petraea'
 
 export default [
   quercusRobur,
-  quercusPetrea
+  quercusPetraea
 ]
